pyproject.toml
[project]
name = "ai-agent-whatsapps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.1.1",
    "litellm>=1.72.1",
]
